+++ public/index.html
@@ -32,6 +32,6 @@
   <!-- Статистика фильмов -->
 </footer>
 
-<script src="../src/main.js"></script>
+<script src="./bundle.js"></script>
 </body>
 </html>
