+++ src/components/footer-statistics.js
@@ -1,0 +1,9 @@
+const createFooterStatisticsTemplate = () => {
+  return (
+    `<section class="footer__statistics">
+      <p>130 291 movies inside</p>
+    </section>`
+  );
+};
+
+export {createFooterStatisticsTemplate};
